backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
dj-database-url==3.0.1
Django==5.1.1
django-admin-datta==1.0.10
django-allauth==65.0.2
django-api-generator==1.0.17
django-colorfield==0.11.0
django-cors-headers==4.7.0
django-countries==7.6.1
django-crispy-forms==2.3
django-debug-toolbar==4.2.0
django-dynamic-datatb==1.0.25
django-filter==25.1
django-mysql==4.12.0
django_celery_results==2.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
fonttools==4.54.1
gunicorn==21.2.0
kiwisolver==1.4.7
kombu==5.5.3
matplotlib==3.9.2
mysqlclient==2.2.7
numpy==2.1.1
packaging==24.1
pandas==2.2.3
pillow==10.2.0
prompt_toolkit==3.0.51
PyJWT==2.10.1
PyMySQL==1.1.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
sorl-thumbnail==12.10.0
sqlparse==0.4.4
str2bool==1.1
typing_extensions==4.10.0
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.6.0
